requirements.txt
python-dotenv==0.20.0
requests==2.27.1
spotipy==2.20.0
tqdm==4.64.0
urllib3==1.26.11
